Move gf_mulx into a shared source file
