+++ gf_common.h
@@ -1,0 +1,1 @@
+void gf_mulx(void *);
